requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
Django==5.2.9
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
dnspython==2.8.0
grpcio==1.76.0
grpcio-tools==1.76.0
gunicorn==23.0.0
mongoengine==0.29.1
packaging==25.0
protobuf==6.33.2
pymongo==4.15.5
python-decouple==3.8
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
